fix npm package issues
The first issue was version mismatches of ethers, so that SignerWithAddress
wasn't matching the expected input of a bunch of functions.
Solution: upgrade to the latest ethers across the board.

Then the issue arose with the `changeEtherBalance` matcher throwing errors
that `getAddress` was not found on objects that worked before.
Solution: wrap EOA in a Wallet object to make it compatible, and stop
using this matcher for contract balances.